+++ infra/audioplayer/impl/src/androidMain/kotlin/app/campfire/audioplayer/impl/content/CoverContentUserComponent.kt
@@ -1,0 +1,10 @@
+package app.campfire.audioplayer.impl.content
+
+import app.campfire.account.api.UrlHydrator
+import app.campfire.core.di.UserScope
+import com.r0adkll.kimchi.annotations.ContributesTo
+
+@ContributesTo(UserScope::class)
+interface CoverContentUserComponent {
+  val urlHydrator: UrlHydrator
+}